pyproject.toml
[project]
name = "emorec"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=23.0.0",
    "fastapi>=0.104.0",
    "httpx>=0.25.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "pillow>=10.0.0",
    "pydantic>=2.4.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "python-dotenv>=1.0.0",
    "ruff>=0.1.0",
    "scikit-learn>=1.3.0",
    "sentence-transformers>=2.2.0",
    "spotipy>=2.23.0",
    "uvicorn[standard]>=0.24.0",
]
